+++ starboard/shared/modular/cobalt_layer_posix_ioctl_abi_wrappers.cc
@@ -1,4 +1,4 @@
-// Copyright 2016 The Cobalt Authors. All Rights Reserved.
+// Copyright 2026 The Cobalt Authors. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,8 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "starboard/thread.h"
+#include <sys/ioctl.h>
 
-SbThreadId SbThreadGetId() {
-  return 0;
+extern "C" {
+
+int __abi_wrap_ioctl_FIONREAD(int fd, int* arg);
+
+int ioctl_FIONREAD(int fd, int* arg) {
+  return __abi_wrap_ioctl_FIONREAD(fd, arg);
 }
+}  // extern "C"