+++ starboard/shared/modular/cobalt_layer_posix_sched_abi_wrappers.cc
@@ -17,9 +17,31 @@
 extern "C" {
 
 int __abi_wrap_sched_getaffinity(pid_t pid, size_t cpusetsize, cpu_set_t* mask);
+int __abi_wrap_sched_getparam(pid_t pid, struct sched_param* param);
+int __abi_wrap_sched_setparam(pid_t pid, const struct sched_param* param);
+int __abi_wrap_sched_getscheduler(pid_t pid);
+int __abi_wrap_sched_setscheduler(pid_t pid,
+                                  int policy,
+                                  const struct sched_param* param);
 
 int sched_getaffinity(pid_t pid, size_t cpusetsize, cpu_set_t* mask) {
   return __abi_wrap_sched_getaffinity(pid, cpusetsize, mask);
+}
+
+int sched_getparam(pid_t pid, struct sched_param* param) {
+  return __abi_wrap_sched_getparam(pid, param);
+}
+
+int sched_setparam(pid_t pid, const struct sched_param* param) {
+  return __abi_wrap_sched_setparam(pid, param);
+}
+
+int sched_getscheduler(pid_t pid) {
+  return __abi_wrap_sched_getscheduler(pid);
+}
+
+int sched_setscheduler(pid_t pid, int policy, const struct sched_param* param) {
+  return __abi_wrap_sched_setscheduler(pid, policy, param);
 }
 
 }  // extern "C"